Refactor skills
