+++ src/features/bumpkins/components/SkillPointsLabel.tsx
